[add] meditation music game password-change lunch-snacks-time-setting & crons calendar-today-yesterday-tomorrow

+++ routes/reminder.routes.js
@@ -6,22 +6,25 @@ const reminderDbService = require("../dbops/reminder.dbops");
 const reminderDbModel = require("../models/reminder.models");
 var ObjectId = require('mongoose').Types.ObjectId;
 
-router.get("/:username", async (req, res) => {
+router.get("/:username/:from/:to", async (req, res) => {
     try {
-        console.log(`/${req.params.username} called`);
-        const todayStart = new Date().setHours(0, 0, 0, 0);
-        const todayEnd = new Date(+todayStart + 86400000);
-        const reminderResp = await reminderDbService.getAllRemindersByOptions(
-            { username: req.params.username, time: { $gte: todayStart, $lt: todayEnd } }
-        );
-        if (reminderResp) {
+        console.log(`/${req.params.username}/${req.params.from}/${req.params.to} called`);
+        if (req.params.username && req.params.from && req.params.to) {
+            const reminderResp = await reminderDbService.getAllRemindersByOptions(
+                { username: req.params.username, time: { $gte: new Date(+req.params.from), $lt: new Date(+req.params.to) } }
+            );
+            if (reminderResp) {
+                return res
+                    .status(200)
+                    .send({ status: true, message: "Reminders found", data: reminderResp });
+            }
             return res
-                .status(200)
-                .send({ status: true, message: "Reminders found", data: reminderResp });
+                .status(404)
+                .send({ status: false, message: "Reminders not found", data: [] });
         }
         return res
-            .status(404)
-            .send({ status: false, message: "Reminders not found", data: [] });
+            .status(400)
+            .send({ status: false, message: "Required fields missing", data: null });
     } catch (e) {
         console.log(e);
         return res.status(500).send(e);
@@ -45,7 +48,7 @@ router.put("/:username", async (req, res) => {
             }
             return res
                 .status(409)
-                .send({ status: false, message: "Reminder not added", data: reminderResp });
+                .send({ status: false, message: "Reminder not added", data: null });
         }
         return res
             .status(400)
@@ -59,15 +62,20 @@ router.put("/:username", async (req, res) => {
 router.delete("/:id", async (req, res) => {
     try {
         console.log(`/${req.params.id} called`);
-        const reminderResp = await reminderDbService.deleteReminderByOptions({ _id: new ObjectId(req.params.id) });
-        if (reminderResp) {
+        if (req.params.id) {
+            const reminderResp = await reminderDbService.deleteReminderByOptions({ _id: new ObjectId(req.params.id) });
+            if (reminderResp) {
+                return res
+                    .status(200)
+                    .send({ status: true, message: "Reminder deleted", data: reminderResp });
+            }
             return res
-                .status(200)
-                .send({ status: true, message: "Reminder deleted", data: reminderResp });
+                .status(409)
+                .send({ status: false, message: "Reminder not deleted", data: null });
         }
         return res
-            .status(404)
-            .send({ status: false, message: "Reminder not deleted", data: null });
+            .status(400)
+            .send({ status: false, message: "Required fields missing", data: null });
     } catch (e) {
         console.log(e);
         return res.status(500).send(e);
